feat(external): moved Pixabay api usage from planner to external
